pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "ai-agent-getting-started"
version = "0.1.0"
description = "A simple AI agent to get started with FastAPI and Python"
authors = ["Daniël Höhle"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
requests = "^2.32.3"
langchain-ollama = "^0.3.3"
langchain-community = "^0.3.24"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
haystack-ai = "^2.14.1"
duckduckgo-api-haystack = "^0.1.15"
gunicorn = "^23.0.0"
uvicorn-worker = "^0.3.0"
langgraph = "^0.4.7"

[[tool.poetry.packages]]
include = "**/*"
from = "src"

[[tool.poetry.source]]
name = "pypi"

[tool.poetry.group.dev.dependencies]
pytest = "^6.2.5"
pytest-cov = "^3.0.0"
pytest-xdist = "^2.4.0"
mypy = "^1.8.0"
types-requests = "^2.26.0"
pre-commit = "^3.7.1"
platformdirs = "^4.2.2"
filelock = "^3.14.0"
distlib = "^0.3.8"

[tool.pytest.ini_options]
addopts = [
    "-v",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=xml",
    "--cov-report=html",
    "--cov-fail-under=70",
    "--cov-context=test"
]
minversion = "6.2.5"
testpaths = [
    "tests"
]
python_paths = [
    "src"
]
